refactor(core): Extract sanitizeSvg and escapeHtml utilities into core
