--- packages/solid/src/internal/escape-html.ts
@@ -1,18 +1,0 @@
-export function escapeHtml(input: string): string {
-  return input.replace(/[&<>"']/g, ch => {
-    switch (ch) {
-      case '&':
-        return '&amp;'
-      case '<':
-        return '&lt;'
-      case '>':
-        return '&gt;'
-      case '"':
-        return '&quot;'
-      case "'":
-        return '&#39;'
-      default:
-        return ch
-    }
-  })
-}
