+++ packages/svelte/src/CircleFlag.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import { codeToEmoji } from '@sankyu/circle-flags-core'
+  import { codeToEmoji, sanitizeSvg } from '@sankyu/circle-flags-core'
 
   /**
    * @deprecated `CircleFlag` is deprecated and not recommended for new code.
@@ -46,52 +46,6 @@
   let loading = $state(true)
 
   const toCssSize = (value: number | string) => (typeof value === 'number' ? `${value}px` : value)
-
-  const sanitizeSvg = (raw: string): string => {
-    if (typeof DOMParser === 'undefined') return raw
-
-    const parser = new DOMParser()
-    const doc = parser.parseFromString(raw, 'image/svg+xml')
-    const svg = doc.documentElement
-
-    if (!svg || doc.querySelector('parsererror')) return raw
-
-    svg.querySelectorAll('script,foreignObject').forEach(el => el.remove())
-
-    svg.querySelectorAll('*').forEach(el => {
-      Array.from(el.attributes).forEach(attr => {
-        const name = attr.name.toLowerCase()
-        const value = attr.value.trim().toLowerCase()
-
-        if (name.startsWith('on')) {
-          try {
-            el.removeAttribute(attr.name)
-          } catch {
-            try {
-              el.setAttribute(attr.name, '')
-            } catch {
-              // Ignore in test / non-DOM environments
-            }
-          }
-          return
-        }
-
-        if ((name === 'href' || name === 'xlink:href') && value.startsWith('javascript:')) {
-          try {
-            el.removeAttribute(attr.name)
-          } catch {
-            try {
-              el.setAttribute(attr.name, '')
-            } catch {
-              // Ignore in test / non-DOM environments
-            }
-          }
-        }
-      })
-    })
-
-    return svg.outerHTML
-  }
 
   $effect(() => {
     if (!finalCode) {
